# simple streaming kernel for the i8mm case