@silvasean you were right, the "max streaming" idea from ruy kernels still helps, even on big OOO cores that "should" be able to make that optimization by themselves.

This is a 2x partial unrolling + pipelining of the inner loop that allows using all of the available registers to maximize load-to-use distance.

It might be counterproductive on cores that care foremost about minimum loop code size such as the Cortex-X1, but I don't have one here to test. Will be a good use case for tuning.

On the 3 core types in the Snapdragon 8gen1 in my Moto Edge x30:

  | Gop/s before | Gop/s after | Gop/s peak | Speedup | %peak before | %peak after
-- | -- | -- | -- | -- | -- | --
Cortex-A510 | 83.4 | 103.0 | 114.24 | 1.24 | 73% | 90%
Cortex-A710 | 307.7 | 308.0 | 319.49 | 1.00 | 96% | 96%
Cortex-X2 | 675.4 | 732.3 | 766.77 | 1.08 | 88% | 96%