@bjacob 2022-09-26
@benvanik the other thing this tells us, given this does 90% of peak on little Cortex-A510 cores, is that we won't need further tuning for those. They look quite different from the A53 and A55 little cores of the past, which required heavy tuning to get above ~60% of peak.